backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
python-multipart==0.0.9

# LLM & Orchestration
langchain>=0.3.0
langchain-ollama>=0.3.0
langgraph>=0.4.0

# RAG & Vector Store
chromadb>=0.6.0

# YouTube
google-api-python-client>=2.100.0
youtube-transcript-api>=1.2.3

# PDF Parsing
PyPDF2>=3.0.0

# MCP Server
mcp[cli]>=1.0.0

# Utilities
pydantic>=2.0.0
httpx>=0.28.0
